requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Queue
redis==5.0.1
rq==1.15.1
apscheduler==3.10.4

# Machine Learning
scikit-learn==1.4.2
xgboost==2.1.1
pandas==2.2.2
numpy<2.0.0,>=1.24.0
joblib==1.3.2

# Vector Search & RAG
azure-search-documents==11.4.0
sentence-transformers==2.2.2
faiss-cpu==1.8.0
huggingface_hub==0.19.4
langchain==0.1.0
langchain-community==0.0.10
langchain-text-splitters==0.0.1
pypdf==3.17.4

# Azure Services
azure-identity==1.17.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.8.1
azure-search-documents==11.4.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & API
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-instrumentation-fastapi==0.57b0
opentelemetry-instrumentation-sqlalchemy==0.57b0
opentelemetry-instrumentation-redis==0.57b0
opentelemetry-exporter-otlp==1.36.0

# Data Processing
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
ruff==0.1.7
black==23.11.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# Container & Deployment
gunicorn==21.2.0
docker==6.1.3



